Added existence check to installations

+++ setup.god
@@ -4,52 +4,11 @@
 #        ----------> https://github.com/ibraheemdev/modern-unix <----------           #
 #                                       |----------> <3U
 
+source ./utils.sh
+
 mkdir -p ~/.flux-capacitor/logs
 curr_tstamp=$(date "+%Y.%m.%d-%H.%M.%S")
 log_file=~/.flux-capacitor/logs/flux-capacitor-$curr_tstamp.log
-
-# install $PACKAGE_NAME 
-# Checks which package manager exist and installs the package using the appropriate one
-install (){
-   if   [ -x "$(command -v apk)" ];       then sudo apk add --no-cache $1
-   elif [ -x "$(command -v apt-get)" ]; then sudo apt-get install $1
-   elif [ -x "$(command -v dnf)" ];     then sudo dnf install $1
-   elif [ -x "$(command -v zypper)" ];  then sudo zypper install $1
-   elif [ -x "$(command -v pacman)" ];  then sudo pacman -S $1
-   else  echo "Sorry, package manager not detected. Please modify the script to run smoothly and share!" >> $log_file
-   fi
-}
-
-# install_deb $DEB_FILE
-install_deb (){
-   if [ -x "$(command -v dpkg)" ];       then sudo dpkg -i $1
-   else  echo "Sorry, deb installer not detected. Please modify the script to run smoothly and share!"  >> $log_file
-   fi
-}
-
-# try_install $PACKAGE_NAME $PACKAGE_URL   [$ARCH flag supported]
-# Checks if the package is already installed, if not, tries to install it from repositories or the given .deb URL
-try_install () {
-   if [ ! -x '$(command -v $1)' ]; then
-      banner $1
-      install $1
-      if [ $? -gt 0 ]
-      then
-         wget $2
-         FILE=$(basename $2)
-         install_deb $FILE 
-         if [ $? -gt 0 ]; then echo "$1 with repo $2 not installed. Please install manually." >> $log_file ; fi
-         rm $FILE 
-      fi
-   fi
-   # WRITE TO CONFIG OR LOG
-}
-
-# banner $TEXT
-# Prints a fantasy LGTBI* friendly banner with the given text
-banner() {
-   toilet -f bigmono9 -F gay $1 ; sleep 0.5 
-}
 
 ARCH=""
 case $(uname -m) in
@@ -87,38 +46,47 @@ try_install lsd https://github.com/Peltoche/lsd/releases/download/0.21.0/lsd_0.2
 if [ $(grep ~/.zshrc -e lsd | wc -c) -eq 0 ] ; then echo "alias l='lsd -lah'" >> ~/.zshrc ; fi
 
 try_install duf https://github.com/muesli/duf/releases/download/v0.8.1/duf_0.8.1_linux_$ARCH.deb
-try_install broot
 
-if [ $? -gt 0 ] && [ "$(uname -o)" = "GNU/Linux" ] && [ "$(uname -m)" = "x86_64" ]  ; then 
-   wget https://dystroy.org/broot/download/x86_64-linux/broot
-   sudo chmod a+x broot
-   sudo mv broot /usr/local/bin/
-else echo "broot not installed. Please install manually." >> $ERRLOG_FILE ; fi
+# try_install broot
+if [ ! "$(command -v mcfly)" ]; then
+   banner "broot"
+   if [ $? -gt 0 ] && [ "$(uname -o)" = "GNU/Linux" ] && [ "$(uname -m)" = "x86_64" ]  ; then 
+      wget https://dystroy.org/broot/download/x86_64-linux/broot
+      sudo chmod a+x broot
+      sudo mv broot /usr/local/bin/
+   else echo "broot not installed. Please install manually." >> $ERRLOG_FILE ; fi
+fi
 
-try_install fd-find https://github.com/sharkdp/fd/releases/download/v8.3.2/fd_8.3.2_$ARCH.deb
+try_install fd-find https://github.com/sharkdp/fd/releases/download/v8.3.2/fd_8.3.2_$ARCH.deb fd
 
-banner "mcfly"
-if [ "$(uname -o)" = "GNU/Linux" ] && [ "$(uname -m)" = "x86_64" ]  ; then 
-   repo="https://github.com/cantino/mcfly/releases/download/v0.5.13/mcfly-v0.5.13-x86_64-unknown-linux-musl.tar.gz"
+if [ ! "$(command -v mcfly)" ]; then
+   banner "mcfly"
+   if [ "$(uname -o)" = "GNU/Linux" ] && [ "$(uname -m)" = "x86_64" ]  ; then 
+      repo="https://github.com/cantino/mcfly/releases/download/v0.5.13/mcfly-v0.5.13-x86_64-unknown-linux-musl.tar.gz"
+      wget $repo
+      tar -zxvf $(basename $repo)
+      sudo chmod a+x mcfly && sudo mv mcfly /usr/local/bin/
+      rm -r $(basename $repo)*
+      if [ $(grep ~/.zshrc -e mcfly | wc -c) -eq 0 ] ; then echo 'eval $(mcfly init zsh)' >> ~/.zshrc ; fi
+      if [ $(grep ~/.bashrc -e mcfly | wc -c) -eq 0 ] ; then echo 'eval $(mcfly init bash)' >> ~/.bashrc ; fi
+   fi 
+fi
+
+
+if [ ! -d ~/.fzf ]; then
+   banner "fzf"
+   git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
+   ~/.fzf/install --all
+fi
+
+if [ ! "$(command -v cheat)" ]; then
+   banner "cheat"
+   mkdir tmp && cd tmp
+   repo="https://github.com/cheat/cheat/releases/download/4.2.3/cheat-linux-$ARCH.gz"
    wget $repo
-   tar -zxvf $(basename $repo)
-   sudo chmod a+x mcfly && sudo mv mcfly /usr/local/bin/
-   rm -r $(basename $repo)*
-   if [ $(grep ~/.zshrc -e mcfly | wc -c) -eq 0 ] ; then echo 'eval $(mcfly init zsh)' >> ~/.zshrc ; fi
-   if [ $(grep ~/.bashrc -e mcfly | wc -c) -eq 0 ] ; then echo 'eval $(mcfly init bash)' >> ~/.bashrc ; fi
- fi 
-
-banner "fzf"
-git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
-~/.fzf/install --all
-
-
-banner "cheat"
-mkdir tmp && cd tmp
-repo="https://github.com/cheat/cheat/releases/download/4.2.3/cheat-linux-$ARCH.gz"
-wget $repo
-gzip -d $(basename $repo) && rm $(basename $repo)
-sudo chmod a+x cheat* && sudo mv cheat* /usr/local/bin/cheat
+   gzip -d $(basename $repo) && rm $(basename $repo)
+   sudo chmod a+x cheat* && sudo mv cheat* /usr/local/bin/cheat
+fi
 
 # banner "exa"
 # exa_repo="https://github.com/ogham/exa/releases/download/v0.10.1/exa-linux-x86_64-v0.10.1.zip"
@@ -133,4 +101,4 @@ sudo chmod a+x cheat* && sudo mv cheat* /usr/local/bin/cheat
 # fi
 
 
-spd-say -p 10 -l ES "¡De super puta madre socio!" 
+# spd-say -p 10 -l ES "¡De super puta madre socio!"
